SWING + ORACLE (not fully implemented yet)

+++ src/Model/Money.java
@@ -2,15 +2,15 @@ package Model;
 
 public class Money {
 
-    private final float quantity;
+    private final double quantity;
     private final Currency currency;
 
-    public Money(float quantity, Currency currency) {
+    public Money(double quantity, Currency currency) {
         this.quantity = quantity;
         this.currency = currency;
     }
 
-    public float getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
